+++ maser/script.py
@@ -29,7 +29,7 @@ from pprint import pformat
 # (define here the global variables)
 logger = logging.getLogger(__name__)
 
-INPUT_DATE = "%Y-%m-%dT%H:%M:%S"
+INPUT_DATE = '%Y-%m-%dT%H:%M:%S'
 
 # ________________ Class Definition __________
 # (If required, define here classes)
@@ -42,19 +42,19 @@ def main():
     parser = argparse.ArgumentParser(add_help=True,
                                      description='MASER4PY application')
     parser.add_argument('--version', action='store_true',
-                        help="Print MASER4PY version")
+                        help='Print MASER4PY version')
     parser.add_argument('-l', '--log-file', nargs=1,
                         default=[None],
-                        help="log file path")
+                        help='log file path')
     parser.add_argument('-Q', '--quiet',
                         action='store_true',
-                        help="Quiet mode")
+                        help='Quiet mode')
     parser.add_argument('-D', '--debug',
                         action='store_true',
-                        help="Debug mode")
+                        help='Debug mode')
 
     # Add maser subparsers
-    subparsers = parser.add_subparsers(dest="maser",
+    subparsers = parser.add_subparsers(dest='maser',
                                        description='maser sub-commands')
 
     # Initializing subparsers
@@ -76,19 +76,19 @@ def main():
     )
 
     if args.version:
-        print("This is MASER4PY V{0}".format(MASER_VERSION))
+        print('This is MASER4PY V{0}'.format(MASER_VERSION))
     elif args.maser is not None:
         # skeletoncdf sub-command
         if 'skeletoncdf' in args.maser:
             skeletons = args.skeletons
             nskt = len(skeletons)
-            logger.info("{0} input file(s) found.".format(nskt))
+            logger.info('{0} input file(s) found.'.format(nskt))
             # Initializing list of bad conversion encountered
             bad_skt = []
             # Loop over the input skeleton files
             for i, skt in enumerate(skeletons):
                 logger.info(
-                    "Executing skeletoncdf for {0}... [{1}/{2}]".format(skt, i + 1, nskt))
+                    'Executing skeletoncdf for {0}... [{1}/{2}]'.format(skt, i + 1, nskt))
                 try:
                     cdf = skeletoncdf(skt,
                                       output_dir=args.output_dir[0],
@@ -97,10 +97,10 @@ def main():
                                       auto_pad=not args.no_auto_pad,
                                       no_cdf=args.no_cdf)
                 except CDFSerializerError as strerror:
-                    logger.exception("SkeletonCDF error has occurred!")
+                    logger.exception('SkeletonCDF error has occurred!')
                     cdf = None
                 except ValueError as strerror:
-                    logger.exception("Value error has occurred!")
+                    logger.exception('Value error has occurred!')
                     cdf = None
                 except:
                     logger.exception('skeletoncdf has failed!')
@@ -109,25 +109,25 @@ def main():
                     if cdf is None:
                         bad_skt.append(skt)
                         logger.error(
-                            "Converting {0} has failed, aborting!".format(skt))
+                            'Converting {0} has failed, aborting!'.format(skt))
                         if not args.force:
                             sys.exit(-1)
 
             if len(bad_skt) > 0:
                 logger.warning(
-                    "Following files have not been converted correctly:")
+                    'Following files have not been converted correctly:')
                 for bad in bad_skt:
                     logger.warning(bad)
         elif 'skeletontable' in args.maser:
             cdfs = args.cdf
             ncdf = len(cdfs)
-            logger.info("{0} input file(s) found.".format(ncdf))
+            logger.info('{0} input file(s) found.'.format(ncdf))
             # Initializing list of bad conversion encountered
             bad_cdf = []
             # Loop over the input CDF files
             for i, cdf in enumerate(cdfs):
                 logger.info(
-                    "Executing skeletontable for {0}... [{1}/{2}]".format(cdf, i + 1, ncdf))
+                    'Executing skeletontable for {0}... [{1}/{2}]'.format(cdf, i + 1, ncdf))
                 skt = skeletontable(cdf,
                                     to_xlsx=args.to_xlsx,
                                     output_dir=args.output_dir[0],
@@ -136,10 +136,10 @@ def main():
                 try:
                     pass
                 except CDFSerializerError as strerror:
-                    logger.error("SkeletonCDF error -- {0}".format(strerror))
+                    logger.error('SkeletonCDF error -- {0}'.format(strerror))
                     cdf = None
                 except ValueError as strerror:
-                    logger.error("Value error -- {0}".format(strerror))
+                    logger.error('Value error -- {0}'.format(strerror))
                     cdf = None
                 except:
                     logger.error(sys.exc_info()[0])
@@ -148,13 +148,13 @@ def main():
                     if cdf is None:
                         bad_cdf.append(cdf)
                         logger.error(
-                            "Converting {0} has failed, aborting!".format(cdf))
+                            'Converting {0} has failed, aborting!'.format(cdf))
                         if not args.force:
                             sys.exit(-1)
 
             if len(bad_cdf) > 0:
                 logger.warning(
-                    "Following files have not been converted correctly:")
+                    'Following files have not been converted correctly:')
                 for bad in bad_cdf:
                     logger.warning(bad)
         # leapsec sub-command
@@ -168,7 +168,7 @@ def main():
             lst = Lstable(file=args.filepath[0])
             if args.date[0] is not None:
                 date = datetime.strptime(args.date[0], INPUT_DATE)
-                print("{0} sec.".format(lst.get_leapsec(date=date)))
+                print('{0} sec.'.format(lst.get_leapsec(date=date)))
             elif args.SHOW_TABLE:
                 print(lst)
             else:
@@ -181,15 +181,17 @@ def main():
                 result = cdf_compare(args.cdf_filepath1, args.cdf_filepath2,
                                      list_ignore_gatt=args.ignore_gatt,
                                      list_ignore_zvar=args.ignore_zvar,
-                                     list_ignore_vatt=args.ignore_vatt)
+                                     list_ignore_vatt=args.ignore_vatt,
+                                     list_numerical_precision=args.precision)
+
                 if result:
-                    logger.info("CDF compare final result :\n %s",
+                    logger.info('CDF compare final result :\n %s',
                                 pformat(result, width=1000))
                 else:
-                    logger.info("CDF compare final result : no differences")
+                    logger.info('CDF compare final result : no differences')
 
             except Exception as err:
-                logger.error("cdf_compare error -- {0}, aborting!".format(err))
+                logger.error('cdf_compare error -- {0}, aborting!'.format(err))
             finally:
                 if result is None:
                     logger.error('CDF compare : Failure !')
@@ -208,15 +210,15 @@ def main():
                              run_cdf_validate=args.run_cdfvalidate)
             except ValidatorException as strerror:
                 logger.error(
-                    "cdf_validator error -- {0}, aborting!".format(strerror))
+                    'cdf_validator error -- {0}, aborting!'.format(strerror))
                 sys.exit(-1)
             except:
-                logger.error("cannot run cdf_validator, aborting!")
+                logger.error('cannot run cdf_validator, aborting!')
                 sys.exit(-1)
         elif 'hfcviewer' in args.maser:
             hfcviewer(**args.__dict__)
         else:
-            print("Unknown maser sub-command")
+            print('Unknown maser sub-command')
             parser.print_help()
 
     else:
@@ -225,5 +227,5 @@ def main():
     # _________________ Main ____________________________
 
 
-if __name__ == "__main__":
+if __name__ == '__main__':
     main()