+++ maser/utils/cdf/cdfcompare/cdf_compare.py
@@ -18,9 +18,9 @@ logger = logging.getLogger(__name__)
 # Checking file
 def checking_file_exist(cdf_file):
     if os.path.isfile(cdf_file) == False:
-        raise FileNotFoundError("%s : does not exist !" % cdf_file)
+        raise FileNotFoundError('%s : does not exist !' % cdf_file)
     if os.access(cdf_file, os.R_OK) == False:
-        raise IOError("%s : is not readable !" % cdf_file)
+        raise IOError('%s : is not readable !' % cdf_file)
 
 
 #
@@ -119,17 +119,17 @@ def compare_global_attributes(global_att1, global_att2, list_ignore_gatt):
     l2 = len(not_match_attribute[1])
 
     if list_ignore_gatt != []:
-        logger.debug("%s Global Attributes to be ignored for comparison : %s", len(list_ignore_gatt),
+        logger.debug('%s Global Attributes to be ignored for comparison : %s', len(list_ignore_gatt),
                        list_ignore_gatt)
     if l1 != 0 or l2 != 0:
-        logger.debug("Global attributes: different")
+        logger.debug('Global attributes: different')
         logger.debug('File 1 : %s%s', str(len(global_att1)), ' global attributes')
         logger.debug("File1's Global Attributes List : %s", list_global_att1)
         logger.debug('File 2 : %s%s', str(len(global_att2)), ' global attributes')
         logger.debug("File2's Global Attributes List : %s", list_global_att2)
-        logger.debug("Not matched global attributes")
-        logger.debug("   File1 : %s - %s", len(not_match_attribute[0]), not_match_attribute[0])
-        logger.debug("   File2 : %s - %s", len(not_match_attribute[1]), not_match_attribute[1])
+        logger.debug('Not matched global attributes')
+        logger.debug('   File1 : %s - %s', len(not_match_attribute[0]), not_match_attribute[0])
+        logger.debug('   File2 : %s - %s', len(not_match_attribute[1]), not_match_attribute[1])
 
         gAttrs['NotMatched'] = not_match_attribute
         # Remove not matched keys from the 2 dictionaries
@@ -150,7 +150,7 @@ def compare_global_attributes(global_att1, global_att2, list_ignore_gatt):
         checking = global_att1 == global_att2
         if checking == False:
             # Not equal !!
-            logger.debug("Global attributes value: different")
+            logger.debug('Global attributes value: different')
             common_att = sorted(list(global_att1.keys()))
 
             DiffValueAttr = {}
@@ -162,35 +162,62 @@ def compare_global_attributes(global_att1, global_att2, list_ignore_gatt):
                         val1 = dd1[0]
                         val2 = dd2[0]
                         logger.debug('** %s', com_att)
-                        logger.debug("   File1 : %s", val1)
-                        logger.debug("   File2 : %s", val2)
+                        logger.debug('   File1 : %s', val1)
+                        logger.debug('   File2 : %s', val2)
 
                         DiffValueAttr[com_att] = [val1, val2]
             if DiffValueAttr:
                 gAttrs['Value'] = DiffValueAttr
-            logger.debug("gAttrs:  %s", pformat(DiffValueAttr, width=1000))
+            logger.debug('gAttrs:  %s', pformat(DiffValueAttr, width=1000))
 
     return gAttrs
 
 
-def compare_z_var(field1, field2, key, shape_diff_dict={}, value_diff_dict={}):
+def precision_dict_from_list(precision_list):
+    precision_dict = {}
+    for key_value in precision_list:
+        key, value = key_value.split(':')
+        precision_dict[key] = float(value)
+    return precision_dict
+
+def compare_z_var(field1, field2, key, dict_numerical_precision={}, shape_diff_dict={}, value_diff_dict={}):
     # check if fields have the same shape
-    data1 = field1[...]
-    data2 = field2[...]
+    if field1.shape:
+        data1 = field1[...]
+    else:
+        # Make sure to have numpy ndarray for data1
+        data1 = np.ndarray(field1[...], dtype=type(field1[...]))
+    if field2.shape:
+        data2 = field2[...]
+    else:
+        # Make sure to have numpy ndarray for data2
+        data2 = np.ndarray((1,), dtype=type(field2[...]))
+        data2[0] = field2[...]
+
+    res = False
+
+    res = False
 
     if data1.shape != data2.shape:
-        logger.debug("%s - array shape is different: %s | %s", key, field1.shape, field2.shape)
+        logger.debug('%s - array shape is different: %s | %s', key, field1.shape, field2.shape)
         shape_diff_dict[key] = [data1.shape, data2.shape]
 
     else:
         differences_mask = data1 != data2
-
         fields_are_different = np.any(differences_mask)
 
         # zVariable's key : check for different values
         if fields_are_different:
-            logger.debug("Different values for zVariable '%s' : %s | %s", key, field1, field2)
-            value_diff_dict[key] = [data1[differences_mask], data2[differences_mask]]
+            if key in dict_numerical_precision.keys():
+                tab_diff = np.isclose(data1, data2, atol=dict_numerical_precision[key])
+                res = np.all(tab_diff)
+                if not res:
+                    logger.debug("Different values for zVariable '%s' : %s | %s", key, field1, field2)
+                    value_diff_dict[key] = [data1[differences_mask], data2[differences_mask]]
+
+            else:
+                logger.debug("Different values for zVariable '%s' : %s | %s", key, field1, field2)
+                value_diff_dict[key] = [data1[differences_mask], data2[differences_mask]]
 
     return shape_diff_dict, value_diff_dict
 
@@ -199,7 +226,7 @@ def compare_v_att(field1, field2, key, key_diff_dict={}, value_diff_dict={}, lis
     v_att_keys_diff = list_differences(field1.attrs, field2.attrs)
 
     if list_ignore_vatt != []:
-        logger.debug("%s Variable Attributes to be ignored for comparison : %s", len(list_ignore_vatt),
+        logger.debug('%s Variable Attributes to be ignored for comparison : %s', len(list_ignore_vatt),
                        list_ignore_vatt)
         for item1 in list_ignore_vatt:
             if item1 in v_att_keys_diff[0]:
@@ -221,12 +248,12 @@ def compare_v_att(field1, field2, key, key_diff_dict={}, value_diff_dict={}, lis
     DiffValue_vAttr = {}
 
     for check_item in common_v_att_keys:
-        logger.debug("%s : %s  |  %s", check_item, vAttrsList1[check_item], vAttrsList2[check_item])
+        logger.debug('%s : %s  |  %s', check_item, vAttrsList1[check_item], vAttrsList2[check_item])
 
         if vAttrsList1[check_item] == vAttrsList2[check_item]:
-            logger.debug("%s : equal", check_item)
+            logger.debug('%s : equal', check_item)
         else:
-            logger.debug("%s : not equal", check_item)
+            logger.debug('%s : not equal', check_item)
             DiffValue_vAttr[check_item] = [vAttrsList1[check_item], vAttrsList2[check_item]]
             value_diff_dict[key] = DiffValue_vAttr
             logger.debug("vAttrs['Value'] : %s", value_diff_dict)
@@ -234,22 +261,22 @@ def compare_v_att(field1, field2, key, key_diff_dict={}, value_diff_dict={}, lis
     return key_diff_dict, value_diff_dict
 
 
-def compare_data(cdf1, cdf2, cdf_keys1, cdf_keys2, list_ignore_zvar=[], list_ignore_vatt=[]):
+def compare_data(cdf1, cdf2, cdf_keys1, cdf_keys2, list_ignore_zvar=[], list_ignore_vatt=[], list_numerical_precision=[]):
     zVars = {}  # store data differences
     vAttrs = {}  # store attribute differences
 
     # Ignored zVariables for comparison
     if list_ignore_zvar != []:
-        logger.debug("%s zVariables to be ignored for comparison : %s", len(list_ignore_zvar), list_ignore_zvar)
+        logger.debug('%s zVariables to be ignored for comparison : %s', len(list_ignore_zvar), list_ignore_zvar)
 
     if len(cdf_keys1) != len(cdf_keys2):
-        logger.debug("Zvariables: different")
+        logger.debug('Zvariables: different')
 
     # ***** Not matched keys *****
     list_diff1, list_diff2 = list_differences(cdf_keys1, cdf_keys2)
 
     if cdf_keys1 == cdf_keys2:
-        logger.debug("Zvariables keys: identical")
+        logger.debug('Zvariables keys: identical')
         same_keys = cdf_keys1
         ordered_common_keys = sorted(list(same_keys))
     else:
@@ -257,17 +284,17 @@ def compare_data(cdf1, cdf2, cdf_keys1, cdf_keys2, list_ignore_zvar=[], list_ign
 
             zVars['Keys'] = [list_diff1, list_diff2]
 
-            logger.debug("NOT MATCHED zVARIABLES :")
+            logger.debug('NOT MATCHED zVARIABLES :')
             for idx, diff_list in enumerate(zVars['Keys']):
-                logger.debug("   File %d : %d - %s", idx + 1, len(diff_list), diff_list)
+                logger.debug('   File %d : %d - %s', idx + 1, len(diff_list), diff_list)
         # ***** Matched keys *****
         same_keys = set(cdf_keys1) & set(cdf_keys2)
 
         # ***** Alphabetical order *****
         ordered_common_keys = sorted(list(same_keys))
 
-    logger.debug("MATCHED zVARIABLES : %d", len(ordered_common_keys))
-    logger.debug("    %s", ordered_common_keys)
+    logger.debug('MATCHED zVARIABLES : %d', len(ordered_common_keys))
+    logger.debug('    %s', ordered_common_keys)
 
     # prepare the dicts to store z_var and v_att diff
     v_att_key_diff_dict = {}
@@ -285,7 +312,12 @@ def compare_data(cdf1, cdf2, cdf_keys1, cdf_keys2, list_ignore_zvar=[], list_ign
         field1 = cdf1.raw_var(key)
         field2 = cdf2.raw_var(key)
 
-        compare_z_var(field1, field2, key, shape_diff_dict=z_var_shape_diff_dict, value_diff_dict=z_var_value_diff_dict)
+        dict_numerical_precision = {}
+        if len(list_numerical_precision) != 0:
+            dict_numerical_precision = precision_dict_from_list(list_numerical_precision)
+
+        compare_z_var(field1, field2, key, dict_numerical_precision=dict_numerical_precision,
+                      shape_diff_dict=z_var_shape_diff_dict, value_diff_dict=z_var_value_diff_dict)
 
         compare_v_att(field1, field2, key,
                       list_ignore_vatt=list_ignore_vatt,
@@ -310,7 +342,7 @@ def compare_data(cdf1, cdf2, cdf_keys1, cdf_keys2, list_ignore_zvar=[], list_ign
 
 
 # Comparing 2 CDF files data
-def cdf_compare(cdf_file1, cdf_file2, list_ignore_gatt=[], list_ignore_zvar=[], list_ignore_vatt=[]):
+def cdf_compare(cdf_file1, cdf_file2, list_ignore_gatt=[], list_ignore_zvar=[], list_ignore_vatt=[], list_numerical_precision=[]):
     logger.debug(' CDF file 1 : %s', cdf_file1)
     logger.debug(' CDF file 2 : %s', cdf_file2)
     checking_file_exist(cdf_file1)
@@ -330,7 +362,8 @@ def cdf_compare(cdf_file1, cdf_file2, list_ignore_gatt=[], list_ignore_zvar=[], 
         dict_result['gAttrs'] = gAttrs
 
     zVars, vAttrs = compare_data(cdf1, cdf2, cdf_keys1, cdf_keys2,
-                                 list_ignore_zvar=list_ignore_zvar, list_ignore_vatt=list_ignore_vatt)
+                                 list_ignore_zvar=list_ignore_zvar, list_ignore_vatt=list_ignore_vatt,
+                                 list_numerical_precision=list_numerical_precision)
 
     if zVars:
         dict_result['zVars'] = zVars
@@ -342,15 +375,15 @@ def cdf_compare(cdf_file1, cdf_file2, list_ignore_gatt=[], list_ignore_zvar=[], 
     cdf2.close()
 
     for key, value in dict_result.items():
-        logger.debug("*°*°* %s *°*°*", key)
+        logger.debug('*°*°* %s *°*°*', key)
         for key1, value1 in dict_result[key].items():
-            logger.debug("     *°* %s : %s", key1, value1)
+            logger.debug('     *°* %s : %s', key1, value1)
 
     # Case of we need to force to ignore some zVariables
     # if forced_ignored_zvar != []:
     #     logger.debug("Forced ignored zVariables (Particular case) : %s", forced_ignored_zvar)
 
-    logger.debug("Return value: %s", pformat(dict_result, width=1000))
+    logger.debug('Return value: %s', pformat(dict_result, width=1000))
     return dict_result
 
 
@@ -361,47 +394,73 @@ def main(cdf_file1, cdf_file2):
         list_ignore_gatt = []
         list_ignore_zvar = []
         list_ignore_vatt = []
+        list_numerical_precision = []
 
-        if "--ignore_gatt" in (list_argv):
-            ind_ignore_gatt = (list_argv).index("--ignore_gatt")
+        if '--ignore_gatt' in (list_argv):
+            ind_ignore_gatt = (list_argv).index('--ignore_gatt')
             list_ignore_gatt = []
             for item in list_argv[ind_ignore_gatt + 1:]:
-                if item.startswith("--ignore_") == False:
+                if (item.startswith("--ignore_") == False and item.startswith("--precision") == False):
                     list_ignore_gatt.append(item)
                 else:
                     break
-            logger.warning("Ignored global attributes list : %s", list_ignore_gatt)
+            logger.warning('Ignored global attributes list : %s', list_ignore_gatt)
         else:
-            logger.debug("No global attributes to be ignored")
+            logger.debug('No global attributes to be ignored')
 
-        if "--ignore_zvar" in (list_argv):
-            ind_ignore_zvar = (list_argv).index("--ignore_zvar")
+        if '--ignore_zvar' in (list_argv):
+            ind_ignore_zvar = (list_argv).index('--ignore_zvar')
             list_ignore_zvar = []
             for item in list_argv[ind_ignore_zvar + 1:]:
-                if item.startswith("--ignore_") == False:
+                if (item.startswith("--ignore_") == False and item.startswith("--precision") == False):
                     list_ignore_zvar.append(item)
                 else:
                     break
-            logger.warning("Ignored zVariables list : %s", list_ignore_zvar)
+            logger.warning('Ignored zVariables list : %s', list_ignore_zvar)
         else:
-            logger.debug("No zVariables to be ignored")
+            logger.debug('No zVariables to be ignored')
 
-        if "--ignore_vatt" in (list_argv):
-            ind_ignore_vatt = (list_argv).index("--ignore_vatt")
+        if '--ignore_vatt' in (list_argv):
+            ind_ignore_vatt = (list_argv).index('--ignore_vatt')
             list_ignore_vatt = []
             for item in list_argv[ind_ignore_vatt + 1:]:
-                if item.startswith("--ignore_") == False:
+                if (item.startswith("--ignore_") == False and item.startswith("--precision") == False):
                     list_ignore_vatt.append(item)
                 else:
                     break
-            logger.warning("Ignored variable attributes list : %s", list_ignore_vatt)
+            logger.warning('Ignored variable attributes list : %s', list_ignore_vatt)
         else:
-            logger.debug("No variable attributes to be ignored")
+            logger.debug('No variable attributes to be ignored')
+
+        if "--precision" in (list_argv):
+            ind_precision_zvar = (list_argv).index("--precision")
+            list_numerical_precision = []
+            for item in list_argv[ind_precision_zvar + 1:]:
+                if (item.startswith("--ignore_") == False and item.startswith("--precision") == False):
+                    list_numerical_precision.append(item)
+                else:
+                    break
+            logger.warning("Numerical precision list : %s", list_numerical_precision)
+        else:
+            logger.debug("No zVariable precision set")
+
+        if "--precision" in (list_argv):
+            ind_precision_zvar = (list_argv).index("--precision")
+            list_numerical_precision = []
+            for item in list_argv[ind_precision_zvar + 1:]:
+                if (item.startswith("--ignore_") == False and item.startswith("--precision") == False):
+                    list_numerical_precision.append(item)
+                else:
+                    break
+            logger.warning("Numerical precision list : %s", list_numerical_precision)
+        else:
+            logger.debug("No zVariable precision set")
 
         result = cdf_compare(cdf_file1, cdf_file2,
                              list_ignore_gatt=list_ignore_gatt,
                              list_ignore_zvar=list_ignore_zvar,
-                             list_ignore_vatt=list_ignore_vatt)
+                             list_ignore_vatt=list_ignore_vatt,
+                             list_numerical_precision=list_numerical_precision)
         logger.info("Final result : %s", pformat(result, width=1000))